+++ tests/test_transform.py
@@ -45,7 +45,7 @@ def test_hankel(htype):                           # 1. fht / 2. hqwe / 3. hquad
 
             # Adjust htarg for fht
             if htype == 'fht':
-                lambd, int_pts = transform.get_spline_values(
+                lambd, int_pts = transform.get_dlf_points(
                         htarg['dlf'], off, htarg['pts_per_dec'])
                 htarg['lambd'] = lambd
                 htarg['int_pts'] = int_pts
@@ -65,7 +65,7 @@ def test_hankel(htype):                           # 1. fht / 2. hqwe / 3. hquad
         if htype == 'hquad':  # Lower atol to ensure convergence
             _, htarg = utils.check_hankel('quad', [1e-8], 0)
         elif htype == 'fht':  # Adjust htarg for fht
-            lambd, int_pts = transform.get_spline_values(
+            lambd, int_pts = transform.get_dlf_points(
                     htarg['dlf'], off, htarg['pts_per_dec'])
             htarg['lambd'] = lambd
             htarg['int_pts'] = int_pts
@@ -93,7 +93,7 @@ def test_hankel(htype):                           # 1. fht / 2. hqwe / 3. hquad
             _, htarg = utils.check_hankel('qwe', [1e-8, '', '', 200, 80, .1,
                                                   1e-6, .1, 1000], 0)
         elif htype == 'fht':  # Adjust htarg for fht
-            lambd, int_pts = transform.get_spline_values(
+            lambd, int_pts = transform.get_dlf_points(
                     htarg['dlf'], off, htarg['pts_per_dec'])
             htarg['lambd'] = lambd
             htarg['int_pts'] = int_pts
@@ -112,7 +112,7 @@ def test_hankel(htype):                           # 1. fht / 2. hqwe / 3. hquad
         # # # 3. Spline; pts_per_dec # # #
         if htype == 'fht':
             _, htarg = utils.check_hankel('fht', ['key_201_2012', 20], 0)
-            lambd, int_pts = transform.get_spline_values(
+            lambd, int_pts = transform.get_dlf_points(
                     htarg['dlf'], off, htarg['pts_per_dec'])
             htarg['lambd'] = lambd
             htarg['int_pts'] = int_pts
@@ -140,7 +140,7 @@ def test_hankel(htype):                           # 1. fht / 2. hqwe / 3. hquad
         elif htype == 'hquad':
             _, htarg = utils.check_hankel('quad', None, 0)
         elif htype == 'fht':
-            lambd, int_pts = transform.get_spline_values(
+            lambd, int_pts = transform.get_dlf_points(
                     htarg['dlf'], off, htarg['pts_per_dec'])
             htarg['lambd'] = lambd
             htarg['int_pts'] = int_pts
@@ -185,8 +185,8 @@ def test_qwe():                                                        # 8. qwe
     dat = DATA['fqwe0'][()]
     tres = DATA['tEM0'][()]
     ftarg = dat['ftarg']
-    tEM, _, _ = transform.qwe(ftarg[0], ftarg[1], ftarg[3], dat['sEM'],
-                              dat['intervals'])
+    tEM, _, _ = transform.qwe(ftarg['rtol'], ftarg['atol'], ftarg['maxint'],
+                              dat['sEM'], dat['intervals'])
     assert_allclose(np.squeeze(-tEM*2/np.pi), tres, rtol=1e-3)
 
     # Hankel
@@ -216,10 +216,10 @@ def test_qwe():                                                        # 8. qwe
     assert_allclose(np.squeeze(fEM), dat['freqres'], rtol=1e-5)
 
 
-def test_get_spline_values():                            # 9. get_spline_values
+def test_get_dlf_points():                                  # 9. get_dlf_points
     # Check one example
     filt = filters.key_81_CosSin_2009()
-    out, new_inp = transform.get_spline_values(filt, np.arange(1, 6), -1)
+    out, new_inp = transform.get_dlf_points(filt, np.arange(1, 6), -1)
     # Expected values
     oout = np.array([[6.70925256e-05, 8.19469958e-05, 1.00090287e-04,
                       1.22250552e-04, 1.49317162e-04, 1.82376393e-04,
@@ -260,20 +260,20 @@ def test_get_spline_values():                            # 9. get_spline_values
 
     # Ensure output dimension
     hfilt = filters.anderson_801_1982()
-    out, _ = transform.get_spline_values(hfilt, np.array([1, 1.1]), -1)
+    out, _ = transform.get_dlf_points(hfilt, np.array([1, 1.1]), -1)
     assert_allclose(out.size, 804)
 
     # Check a hypothetical short filter, with small pts_per_dec, and ensure
     # at least four points are returned
     filt = filters.DigitalFilter('shortest')
     filt.base = np.array([1., 1.1])
-    out, new_inp = transform.get_spline_values(filt, np.array([1.]), 1)
+    out, new_inp = transform.get_dlf_points(filt, np.array([1.]), 1)
     assert_allclose(out.size, 4)
 
     # Check standard example
     ffilt = filters.key_81_CosSin_2009()
     inp = np.arange(1, 6)
-    out, new_inp = transform.get_spline_values(ffilt, inp, 0)
+    out, new_inp = transform.get_dlf_points(ffilt, inp, 0)
     assert_allclose(inp, new_inp)
     assert_allclose(out, ffilt.base/inp[:, None])
 
@@ -411,7 +411,7 @@ def test_dlf():                                                       # 10. dlf
         # # # 1. Spline; One angle # # #
 
         # fht calculation
-        lambd, _ = transform.get_spline_values(dlf, off, pts_per_dec)
+        lambd, _ = transform.get_dlf_points(dlf, off, pts_per_dec)
         PJ1 = kernel.wavenumber(zsrc, zrec, lsrc, lrec, depth, etaH, etaV,
                                 zetaH, zetaV, lambd, ab, xdirect, msrc, mrec)
 
@@ -428,7 +428,7 @@ def test_dlf():                                                       # 10. dlf
         off, angle = utils.get_off_ang(src, rec, nsrc, nrec, 0)
 
         # fht calculation
-        lambd, _ = transform.get_spline_values(dlf, off, -1)
+        lambd, _ = transform.get_dlf_points(dlf, off, -1)
         PJ2 = kernel.wavenumber(zsrc, zrec, lsrc, lrec, depth, etaH, etaV,
                                 zetaH, zetaV, lambd, ab, xdirect, msrc, mrec)
         factAng = kernel.angle_factor(angle, ab, msrc, mrec)
@@ -448,7 +448,7 @@ def test_dlf():                                                       # 10. dlf
         off, angle = utils.get_off_ang(src, rec, nsrc, nrec, 0)
 
         # fht calculation
-        lambd, _ = transform.get_spline_values(dlf, off, -1)
+        lambd, _ = transform.get_dlf_points(dlf, off, -1)
         PJ2 = kernel.wavenumber(zsrc, zrec, lsrc, lrec, depth, etaH, etaV,
                                 zetaH, zetaV, lambd, ab, xdirect, msrc, mrec)
         factAng = kernel.angle_factor(angle, ab, msrc, mrec)
@@ -464,7 +464,7 @@ def test_dlf():                                                       # 10. dlf
 
         # # # 3. Spline; Multi angle # # #
 
-        lambd, _ = transform.get_spline_values(dlf, off, 30)
+        lambd, _ = transform.get_dlf_points(dlf, off, 30)
         # fht calculation
         PJ3 = kernel.wavenumber(zsrc, zrec, lsrc, lrec, depth, etaH, etaV,
                                 zetaH, zetaV, lambd, ab, xdirect, msrc, mrec)